split: extract backend guards and fleet checks from test_doctor_migration
Extract backend guard pattern tests and fleet checks into focused companion files:
- test_doctor_backend_guards.py: backend guards + process state + run_doctor wiring
- test_doctor_fleet_checks.py: Group M ambient/infra/campaign + Group N feature gates

Also extract lifecycle tests from test_update_checks_fetch.py:
- test_update_checks_lifecycle.py: UC-11 epoch invalidation, UC-12 transitions, T1/T2/T6 sequence

Updated tests/cli/CLAUDE.md, tests/arch/test_feature_markers.py registry, and
split guard files to reflect the new structure.
